requirements.txt
asgiref==3.3.1
autopep8==1.5.6
certifi==2020.12.5
Django==3.1.7
djangorestframework==3.12.4
nose==1.3.7
ply==3.11
pycodestyle==2.7.0
Pyomo==5.7.3
pytz==2021.1
PyUtilib==6.0.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
